+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 - When no file path is provided, the `break` debugging command defaults to the file the debugger was opened on, not the current file.
+- `assign` commands in the debugger can now be undone.
 
 
 ## [0.6.0] - 2019-02-10